requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.3
absl-py=2.2.2=pypi_0
addict=2.4.0=pyhd8ed1ab_2
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.16=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
aom=3.6.0=h313beb8_0
appnope=0.1.4=pyhd8ed1ab_1
argcomplete=3.6.0=pypi_0
arrow-cpp=19.0.0=h0b7d223_2
asttokens=3.0.0=pyhd8ed1ab_1
astunparse=1.6.3=pypi_0
attrs=25.3.0=pypi_0
aws-c-auth=0.6.19=h80987f9_0
aws-c-cal=0.5.20=h80987f9_0
aws-c-common=0.8.5=h80987f9_0
aws-c-compression=0.2.16=h80987f9_0
aws-c-event-stream=0.2.15=h313beb8_0
aws-c-http=0.6.25=h80987f9_0
aws-c-io=0.13.10=h80987f9_0
aws-c-mqtt=0.7.13=h80987f9_0
aws-c-s3=0.1.51=h80987f9_0
aws-c-sdkutils=0.1.6=h80987f9_0
aws-checksums=0.1.13=h80987f9_0
aws-crt-cpp=0.18.16=h313beb8_0
aws-sdk-cpp=1.11.212=hdd7fb2f_0
blas=1.0=openblas
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.4.2=py311hb9f6ed7_0
brotli-python=1.0.9=py311h313beb8_9
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2025.2.25=hca03da5_0
cachetools=5.5.2=pypi_0
cairo=1.16.0=h302bd0f_5
certifi=2025.1.31=py311hca03da5_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=py311h48ca7d4_0
cryptography=44.0.1=pypi_0
curl=8.12.1=h3e2b118_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h9131b1a_1
datasets=3.5.0=pypi_0
dav1d=1.2.1=h80987f9_0
debugpy=1.8.11=py311h313beb8_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
dill=0.3.8=py311hca03da5_0
duckdb=1.2.2=pypi_0
duckdb-engine=0.13.6=pypi_0
eigen=3.4.0=h48ca7d4_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h313beb8_0
fastapi=0.115.12=pypi_0
fastdup=2.20=pypi_0
ffmpeg=6.1.1=h7693a39_3
filelock=3.18.0=pypi_0
flatbuffers=25.2.10=pypi_0
fontconfig=2.14.1=h6402c1e_3
fonttools=4.57.0=pypi_0
freetype=2.12.1=h1192e45_0
fribidi=1.0.10=h1a28f6b_0
frozenlist=1.5.0=py311h80987f9_0
fsspec=2025.3.0=pypi_0
gast=0.6.0=pypi_0
gdk-pixbuf=2.42.10=h969f7ef_2
gettext=0.21.0=hbdbcc25_2
gflags=2.2.2=h313beb8_1
giflib=5.2.2=h93a5062_0
git=2.45.2=pl5340hae58ea7_1
glib=2.78.4=h313beb8_0
glib-tools=2.78.4=h313beb8_0
glog=0.5.0=h313beb8_1
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py311h5c1b81f_0
google-auth=2.29.0=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=hc377ac9_1
greenlet=3.1.1=pypi_0
grpcio=1.71.0=py311h313beb8_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h11=0.14.0=pypi_0
h5py=3.13.0=pypi_0
harfbuzz=10.2.0=h34cd0bf_0
hdf5=1.14.5=hd77251f_2
httpcore=1.0.8=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
icu=73.1=h313beb8_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=9.0.2=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=py311hca03da5_0
joblib=1.4.2=pypi_0
jpeg=9e=h80987f9_3
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
keras=3.9.2=pypi_0
kiwisolver=1.4.8=py311h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
lcms2=2.16=he93ba84_0
leptonica=1.82.0=hdedf53e_3
lerc=4.0.0=h313beb8_0
libabseil=20250127.0=cxx17_h313beb8_0
libarchive=3.7.7=h8f13d7a_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h80987f9_9
libbrotlidec=1.0.9=h80987f9_9
libbrotlienc=1.0.9=h80987f9_9
libclang=18.1.1=pypi_0
libcurl=8.12.1=hde089ae_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.22=h80987f9_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.78.4=h0a96307_0
libgrpc=1.71.0=h62f6fdd_0
libiconv=1.16=h80987f9_3
libjpeg-turbo=2.0.0=h1a28f6b_0
libnghttp2=1.57.0=h62f6fdd_0
libogg=1.3.5=h1a28f6b_1
libopenblas=0.3.21=h269037a_0
libopus=1.3.1=h80987f9_1
libpng=1.6.39=h80987f9_0
libpq=17.4=h02f6b3c_0
libprotobuf=5.29.3=h9f9f828_0
libre2-11=2024.07.02=h313beb8_0
librsvg=2.56.3=h79aa576_1
libsodium=1.0.18=h27ca646_1
libssh2=1.11.1=h3e2b118_0
libtheora=1.1.1=h1a28f6b_3
libthrift=0.15.0=h73c2103_2
libtiff=4.7.0=h91aec0a_0
libvorbis=1.3.7=h1a28f6b_0
libvpx=1.13.1=h313beb8_0
libwebp=1.3.2=hf40cdb4_1
libwebp-base=1.3.2=h80987f9_1
libxml2=2.13.5=h0b34f26_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py311h80987f9_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
ml-dtypes=0.5.1=pypi_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py311hca03da5_0
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
mysql=8.4.0=h065ec36_2
namex=0.0.8=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=py311hca03da5_0
numexpr=2.10.1=py311h5d9532f_0
numpy=2.1.3=pypi_0
numpy-base=2.2.4=py311hb71ad10_0
opencv=4.10.0=py311he4f9b47_5
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openh264=2.1.1=h98b2900_0
openjpeg=2.5.2=hba36e21_1
openldap=2.6.4=he7ef289_0
openssl=3.4.1=h81ee809_0
opt-einsum=3.4.0=pypi_0
optree=0.15.0=pypi_0
orc=2.1.1=h55d209b_0
packaging=24.2=py311hca03da5_0
pandas=2.2.3=py311hcf29cfe_0
pango=1.50.7=h8ab3420_1
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.42=hb066dcc_1
perl=5.38.2=0_h80987f9_perl5
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=pypi_0
pillow-heif=0.22.0=pypi_0
pip=25.0=py311hca03da5_0
pipx=1.7.1=pypi_0
pixman=0.40.0=h1a28f6b_0
platformdirs=4.3.6=pypi_0
polars=0.20.0=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.3.1=pypi_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pyopenssl=25.0.0=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.11=hb885b13_0
python-dateutil=2.9.0post0=py311hca03da5_2
python-multipart=0.0.20=pypi_0
python-xxhash=3.5.0=py311h80987f9_0
pytorch=2.6.0.dev20241112=py3.11_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=py311h80987f9_0
pyzmq=26.2.0=py311h313beb8_0
qtbase=6.7.3=hb2d0045_0
re2=2024.07.02=h48ca7d4_0
readline=8.2=h1a28f6b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=py311h80987f9_0
requests=2.32.3=py311hca03da5_1
rich=14.0.0=pypi_0
rpds-py=0.23.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.3=py311hc2e03bc_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=py311hca03da5_2
sentry-sdk=2.26.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=78.1.0=pypi_0
six=1.17.0=pypi_0
snappy=1.2.1=h313beb8_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h80987f9_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.46.2=pypi_0
starlette-prometheus=0.10.0=pypi_0
supervision=0.25.1=pypi_0
sympy=1.13.1=pypi_0
tensorboard=2.19.0=py311hca03da5_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.19.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=3.0.1=pypi_0
tesseract=5.2.0=h313beb8_2
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.21.1=pypi_0
torch=2.6.0=pypi_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.5.0.dev20241118=py311_cpu
torchvision=0.21.0=pypi_0
tornado=6.4.2=py311h80987f9_0
tqdm=4.66.3=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.51.3=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
unicodedata2=15.1.0=py311h80987f9_1
urllib3=2.4.0=pypi_0
userpath=1.9.2=pypi_0
utf8proc=2.6.1=h80987f9_1
uvicorn=0.34.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.1.3=py311hca03da5_0
wheel=0.45.1=py311hca03da5_0
wrapt=1.17.2=pypi_0
xxhash=0.8.0=h1a28f6b_3
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.19.0=pypi_0
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h18a0788_1
zstd=1.5.6=hfb09047_0
